[ENTESB-4886] R patch for standalone child is reset, not cherry-pick
(cherry picked from commit 9556341f7271efcfcdf4b1378730ba8806bf9618)

+++ patch/patch-core/src/main/java/io/fabric8/patch/impl/ServiceImpl.java
@@ -136,8 +136,13 @@ public class ServiceImpl implements Service {
             if (dir != null) {
                 patchDir = new File(dir);
             } else {
-                // only now fallback to datafile of system bundle
-                patchDir = this.bundleContext.getDataFile("patches");
+                if (patchManagement.isStandaloneChild()) {
+                    patchDir = new File(System.getProperty("karaf.home"), "patches");
+                }
+                if (patchDir == null) {
+                    // only now fallback to datafile of system bundle
+                    patchDir = this.bundleContext.getDataFile("patches");
+                }
             }
         }
         if (!patchDir.isDirectory()) {
@@ -274,8 +279,19 @@ public class ServiceImpl implements Service {
                     patchData.getId(),
                     what == Pending.ROLLUP_INSTALLATION ? "installed" : "rolled back");
             if (what == Pending.ROLLUP_ROLLBACK) {
-                File file = new File(patchDir, patchData.getId() + ".patch.result");
-                file.delete();
+                List<String> bases = patch.getResult().getKarafBases();
+                for (Iterator<String> iterator = bases.iterator(); iterator.hasNext(); ) {
+                    String s = iterator.next();
+                    if (s.startsWith(System.getProperty("karaf.name"))) {
+                        iterator.remove();
+                    }
+                }
+                patch.getResult().setPending(null);
+                patch.getResult().store();
+                if (patch.getResult().getKarafBases().size() == 0) {
+                    File file = new File(patchDir, patchData.getId() + ".patch.result");
+                    file.delete();
+                }
             }
         }
     }
@@ -431,8 +447,15 @@ public class ServiceImpl implements Service {
                 }
 
                 // prepare patch result before doing runtime changes
-                PatchResult result = new PatchResult(patch.getPatchData(), simulate, System.currentTimeMillis(),
-                        bundleUpdatesInThisPatch, featureUpdatesInThisPatch);
+                PatchResult result = null;
+                if (patch.getResult() != null) {
+                    result = patch.getResult();
+                } else {
+                    result = new PatchResult(patch.getPatchData(), simulate, System.currentTimeMillis(),
+                            bundleUpdatesInThisPatch, featureUpdatesInThisPatch);
+                }
+                result.getKarafBases().add(String.format("%s | %s",
+                        System.getProperty("karaf.name"), System.getProperty("karaf.base")));
                 results.put(patch.getPatchData().getId(), result);
             }
